+++ NoveltyProducer/Technican.py
@@ -4,6 +4,14 @@ from NoveltyProducer.Generator import Generator
 from datetime import datetime
 import json
 
+class InvalidInputTypeError(Exception):
+    """The InvalidInputTypeError is raised whenever a specific input of a specific function has an invalid/unexpected type."""
+    pass
+
+class InvalidInputValueError(Exception):
+    """The InvalidInputValueError is raised whenever a specific input of a specific function has an invalid/unexpected value."""
+    pass
+
 class Technican:
     """Class to manage multiple generator instances at once.
     """
@@ -12,7 +20,18 @@ class Technican:
         """
         # store list locally
         self.generators = generators_
-        
+        # check input types
+        self._check_input()
+            
+    def _check_input(self):
+        """ Check the given input data for type.
+        """
+        # generators_
+        if not isinstance(self.generators, dict):
+            raise InvalidInputTypeError("The parameter generators_ is type %s but should be a of type dict." % type(self.generators))
+        if not all(isinstance(gen, Generator) for key, gen in self.generators.items()):
+            raise InvalidInputValueError("Not all values of generators are an instance of class NoveltyProducer.Generator.")
+            
     def _get_overall_output(self, name_, time_=None):
         """Get the combined output of all generators.